Update index.html
The page still fully works. I think this is enough of a user-friendly interface. Please let me know if otherwise. :)
This page is slightly different now. I managed to make it work where the options don't shift down when the guesses-table is shown, and the page changes dimensions when wanted, e.g. the opts column is right below the data-table, no matter how many points are in the game scatterplot. There are also axis labels and titles for both graphs.